requirements.txt
# Core dependencies
fastapi==0.68.1
uvicorn==0.15.0
PyGithub==1.55
python-dotenv==0.19.0
requests==2.26.0
mcp-server==0.1.0
httpx==0.24.1
pydantic==1.10.12

# Testing dependencies
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-env==1.0.1
pytest-mock==3.11.1
pytest-benchmark==4.0.0
httpx==0.24.1
psutil==5.9.5
aioresponses==0.7.4
responses==0.23.1
